+++ src/main/ml-modules/base/root/config/userFacingConfig.mjs
@@ -113,7 +113,7 @@ const searchTermText = {
     recordType: {
       label: 'Person or Group',
       helpText:
-        'Search only records that are a Person or search only for records that are a Group',
+        'Search for records categorized as either a "Person" or a "Group".',
     },
     startAt: {
       label: 'Born/Formed At',
@@ -251,6 +251,11 @@ const searchTermText = {
       helpText:
         'Search for Concept terms that describe the Occupation of the specified People & Groups.',
     },
+    recordType: {
+      label: 'Concept Class',
+      helpText:
+        'Search for records categorized as a "Measurement Unit", "Language", "Material", "Currency" or "General Concept".',
+    },
     subjectOfConcept: {
       label: 'Subject Of',
       helpText:
@@ -300,6 +305,11 @@ const searchTermText = {
       label: 'Name',
       helpText:
         'Enter term(s) to be found within the title or name of the Event. "AND", "OR", and "-" do not have special meaning in Advanced Search as they do in Simple Search. Instead use multiple fields connected with "have All of", "have Any of", and "have None of" respectively.',
+    },
+    recordType: {
+      label: 'Event Class',
+      helpText:
+        'Search for records categorized as either a "Period" or an "Activity".',
     },
     startDate: {
       label: 'Start Date',
@@ -410,6 +420,11 @@ const searchTermText = {
       label: 'Created Using Technique',
       helpText:
         'Search for Objects that were created using the specified Technique.',
+    },
+    recordType: {
+      label: 'Object Class',
+      helpText:
+        'Search for records categorized as either a "Physical Object" or a "Digital Object".',
     },
     text: {
       label: 'Anywhere',
@@ -637,6 +652,11 @@ const searchTermText = {
       label: 'Published Date',
       helpText:
         'Search for Works by the date on which they were published. Note the difference with Creation Date, which is when the work was originally conceived by its creator.',
+    },
+    recordType: {
+      label: 'Work Class',
+      helpText:
+        'Search for records categorized as either a "Visual Work" or a "Textual Work".',
     },
     text: {
       label: 'Anywhere',