+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All changes to the MarkLogic (backend) portion of LUX capable of impacting the r
 ## Unreleased
 
 ### Added
+- Added facets and search terms for record types for Objects, Works, Concepts, and Events ([#138](https://github.com/project-lux/lux-marklogic/issues/138),[#139](https://github.com/project-lux/lux-marklogic/issues/139),[#140](https://github.com/project-lux/lux-marklogic/issues/140),[#141](https://github.com/project-lux/lux-marklogic/issues/141))
 
 ### Changed