+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All changes to the MarkLogic (backend) portion of LUX capable of impacting the r
 
 ### Added
 - Added facets and search terms for record types for Objects, Works, Concepts, and Events ([#138](https://github.com/project-lux/lux-marklogic/issues/138),[#139](https://github.com/project-lux/lux-marklogic/issues/139),[#140](https://github.com/project-lux/lux-marklogic/issues/140),[#141](https://github.com/project-lux/lux-marklogic/issues/141))
+- Added search terms for creation/production influenced by an agent ([#65](https://github.com/project-lux/lux-marklogic/issues/65))
 
 ### Changed